Extract cart serialization into a separate service.
