+++ src/Modules/OrchardCore.Commerce/Abstractions/IShoppingCartHelpers.cs
@@ -1,6 +1,6 @@
 using OrchardCore.Commerce.Models;
+using OrchardCore.Commerce.MoneyDataType;
 using OrchardCore.Commerce.ViewModels;
-using OrchardCore.ContentManagement.Metadata.Models;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
@@ -17,35 +17,21 @@ public interface IShoppingCartHelpers
     ShoppingCartLineViewModel GetExistingLine(ShoppingCartViewModel cart, ShoppingCartLineViewModel line);
 
     /// <summary>
-    /// Builds the <see cref="ShoppingCart"/> from the provided update model.
-    /// </summary>
-    Task<ShoppingCart> ParseCartAsync(ShoppingCartUpdateModel cart);
-
-    /// <summary>
-    /// Builds a single <see cref="ShoppingCartItem"/> from the provided update model.
-    /// </summary>
-    Task<ShoppingCartItem> ParseCartLineAsync(ShoppingCartLineUpdateModel line);
-
-    /// <summary>
-    /// Returns a set of attribute values using the available <see cref="IProductAttributeProvider"/>s.
-    /// </summary>
-    ISet<IProductAttributeValue> ParseAttributes(ShoppingCartLineUpdateModel line, ContentTypeDefinition type);
-
-    /// <summary>
-    /// Returns a JSON serialized <see langword="string"/> of <paramref name="cart"/>.
-    /// </summary>
-    Task<string> SerializeAsync(ShoppingCart cart);
-
-    /// <summary>
-    /// Returns a deserialized object from JSON string <paramref name="serializedCart"/>.
-    /// </summary>
-    Task<ShoppingCart> DeserializeAsync(string serializedCart);
-
-    /// <summary>
     /// Validates and updates the <paramref name="parsedLine"/>. If fails sends notification and returns <see
     /// langword="null"/>.
     /// </summary>
     Task<ShoppingCartItem> ValidateParsedCartLineAsync(
         ShoppingCartLineUpdateModel line,
         ShoppingCartItem parsedLine);
+
+    /// <summary>
+    /// Calculate the total value in the cart. All prices must be of a single currency.
+    /// </summary>
+    /// <returns>The total value of the items in the cart, or <see langword="null" /> if the cart is empty.</returns>
+    Task<Amount?> CalculateSingleCurrencyTotalAsync();
+
+    /// <summary>
+    /// Groups the line items in the cart by currency and returns the value by currency code.
+    /// </summary>
+    Task<IDictionary<string, Amount>> CalculateMultipleCurrencyTotalsAsync();
 }
